Re-packages 0.04 with an updated Changes file.  Updated the mkupload script to check for this next time.
git-svn-id: file:///Users/ssmith/svn-rep/rmi@148 f4b04360-953d-48bb-9aa4-7cb10f2341d2

+++ mkupload.pl
@@ -25,10 +25,14 @@ unless ($mod_ver->numify == $qver) {
     die "version mismatch!";
 }
 
+unless (grep { /$ver/ } `cat trunk/Changes`) {
+    die "no version info for $ver in the Changes file";
+}
+
 print "svn cp trunk releases/${mod}-${ver}\n";
 print "svn export releases/${mod}-${ver} uploads/${mod}-${ver}\n";
 print "cp -r uploads/${mod}-${ver} x\n";
 print "cd x; perl Makefile.PL; make; make test; cd ..; rm -rf x\n";
 print "cd uploads; tar -cvf ${mod}-${ver}.tar ${mod}-${ver}; gzip --best ${mod}-${ver}.tar; rm -rf ${mod}-${ver}; cd ..\n";
-
+print "svn add uploads/${mod}-${ver}.tar.gz\n";
